cloud_app_demo/lab4/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Tony", email = "gibbons_tony@berkeley.edu"}
]
readme = "README.md"
requires-python = "~3.11"
dependencies = [
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "scikit-learn (==1.5.2)",
    "joblib (==1.4.2)",
    "pydantic (>=2.10.6,<3.0.0)",
    "fastapi-cache2[redis] (==0.1.9)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.9.0"

